Add files via upload
This scripts uses python module "Selenium" and a Chrome webdriver to log into your online **Mofid** Website and buy a predefined stock at a predefined price at a predefined time.
Just run it and it'll ask for your username, password, stock that you want to buy, how many of it you want, the price that you considered buying that stock on, the duration of sending requests, the time on which you want the script to start sending orders. then it will open the site and fills the username and password section and waits for 4 seconds for you to jot down the captcha. then it sends and Enter command itself to login and then you have 6 seconds to close all the prompts off the page to make it neat. then just wait for it to do its job.

caution: the webpage shouldn't ask for any additional confirmation for sending the orders. so go to settings and turn it off in case it's not.